refactor: dto 수정과 역할 분배
- 리턴 DTO에 대해서도 명확하게 명세하도록 수정
- 각각의 API의 리턴 형태를 명확하게 변경
- 엔티티로부터 필요한 데이터만 뽑아올 수 있게 하는 별도의 정적 함수 제공
