+++ backend/src/room/dto/create-room.dto.ts
@@ -8,9 +8,14 @@ import {
     Max,
     Min,
 } from "class-validator";
-import { RoomStatus } from "@/room/room.entity";
+import { Connection, RoomStatus } from "@/room/domain/room";
+import { Question } from "@/question-list/entity/question.entity";
 
 export class CreateRoomDto {
+    private static MAX_CATEGORY_SIZE = 3;
+    private static MIN_CATEGORY_SIZE = 1;
+    private static MAX_PARTICIPANT_SIZE = 5;
+    private static MIN_PARTICIPANT_SIZE = 1;
     @IsNotEmpty()
     title: string;
 
@@ -24,45 +29,32 @@ export class CreateRoomDto {
 
     @IsNotEmpty()
     @IsArray()
-    @ArrayMaxSize(3)
-    @ArrayMinSize(1)
+    @ArrayMaxSize(CreateRoomDto.MAX_CATEGORY_SIZE)
+    @ArrayMinSize(CreateRoomDto.MIN_CATEGORY_SIZE)
     category: string[];
 
     @IsNumber()
-    @Min(1)
-    @Max(5)
+    @Min(CreateRoomDto.MIN_PARTICIPANT_SIZE)
+    @Max(CreateRoomDto.MAX_PARTICIPANT_SIZE)
     maxParticipants: number;
 
     @IsNumber()
     questionListId: number;
 }
 
-export class CreateRoomInternalDto {
-    @IsNotEmpty()
+export interface CreateRoomResponseDto {
     title: string;
-
-    @IsEnum(RoomStatus, {
-        message: "Status must be either PUBLIC or PRIVATE",
-    })
     status: RoomStatus;
-
-    @IsNotEmpty()
     nickname: string;
-
-    @IsNotEmpty()
-    @IsArray()
-    @ArrayMaxSize(3)
-    @ArrayMinSize(1)
-    category: string[];
-
-    @IsNotEmpty()
-    socketId: string;
-
-    @IsNumber()
-    @Min(1)
-    @Max(5)
     maxParticipants: number;
-
-    @IsNumber()
+    category: string[];
     questionListId: number;
+    socketId: string;
+    id: string;
+    inProgress: boolean;
+    connectionMap: Record<string, Connection>;
+    participants: number;
+    questionListContents: Question[];
+    createdAt: number;
+    host: Connection;
 }
