+++ backend/src/room/dto/index.ts
@@ -1,0 +1,7 @@
+export * from "./create-room.dto";
+export * from "./finish-room.dto";
+export * from "./join-room.dto";
+export * from "./move-index.dto";
+export * from "./reaction.dto";
+export * from "./room-id.dto";
+export * from "./room-list.dto";
