+++ backend/src/room/dto/join-room.dto.ts
@@ -1,4 +1,6 @@
 import { IsNotEmpty } from "class-validator";
+import { Question } from "@/question-list/entity/question.entity";
+import { Connection, RoomStatus } from "@/room/domain/room";
 
 export class JoinRoomDto {
     @IsNotEmpty()
@@ -8,13 +10,19 @@ export class JoinRoomDto {
     nickname: string;
 }
 
-export class JoinRoomInternalDto {
-    @IsNotEmpty()
-    roomId: string;
-
-    @IsNotEmpty()
-    socketId: string;
-
-    @IsNotEmpty()
-    nickname: string;
+export interface JoinRoomResponseDto {
+    category: string[];
+    inProgress: boolean;
+    connectionMap: Record<string, Connection>;
+    createdAt: number;
+    currentIndex: number;
+    maxQuestionListLength: number;
+    questionListId: number;
+    host: Connection;
+    participants: number;
+    maxParticipants: number;
+    status: RoomStatus;
+    title: string;
+    id: string;
+    questionListContents: Question[];
 }
