+++ backend/src/room/dto/room-list.dto.ts
@@ -1,30 +1,14 @@
-import { Connection, RoomStatus } from "@/room/room.entity";
-import { Max, Min } from "class-validator";
+import { Connection, RoomStatus } from "@/room/domain/room";
 
-export class RoomList {
+export class RoomListResponseDto {
     createdAt: number;
-
     host: Connection;
-
-    @Max(5)
-    @Min(1)
     maxParticipants: number;
-
     status: RoomStatus;
-
     title: string;
-
     id: string;
-
     category: string[];
-
     inProgress: boolean;
-
     questionListTitle: string;
-
-    @Max(5)
-    @Min(1)
     participants: number;
 }
-
-export type RoomListResponseDto = RoomList[];
